scripts/curation/requirements.txt
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
readability-lxml>=0.8.1
yt-dlp>=2024.1.0
Jinja2>=3.1.0
google-generativeai>=0.8.0
lxml>=5.0.0
python-slugify>=8.0.0
